+++ .gitattributes
@@ -1,1 +1,1 @@
-fsspec_reference_maker/_version.py export-subst
+kerchunk/_version.py export-subst
