+++ .github/workflows/default.yml
@@ -42,7 +42,7 @@ jobs:
     # ===============================
     - name: Commit documentation changes
       run: |
-        git clone https://github.com/intake/fsspec-reference-maker.git --branch gh-pages --single-branch gh-pages
+        git clone https://github.com/fsspec/kerchunk.git --branch gh-pages --single-branch gh-pages
         cp -r docs/_build/html/* gh-pages/
         cd gh-pages
         touch .nojekyll
