+++ MANIFEST.in
@@ -1,5 +1,5 @@
 include versioneer.py
-include fsspec_reference_maker/_version.py
+include kerchunk/_version.py
 include requirements.txt
 include setup.cfg
 include setup.py
