+++ docs/Makefile
@@ -5,7 +5,7 @@
 # from the environment for the first two.
 SPHINXOPTS    =
 SPHINXBUILD   = sphinx-build
-SPHINXPROJ    = fsspec-reference-maker
+SPHINXPROJ    = kerchunk
 SOURCEDIR     = source
 BUILDDIR      = build
