+++ kerchunk/__init__.py
@@ -1,0 +1,7 @@
+from ._version import get_versions
+
+__version__ = get_versions()["version"]
+del get_versions
+
+
+__all__ = ["__version__"]
