+++ setup.cfg
@@ -4,8 +4,8 @@ long_description: file: README.rst
 [versioneer]
 VCS = git
 style = pep440
-versionfile_source = fsspec_reference_maker/_version.py
-versionfile_build = fsspec_reference_maker/_version.py
+versionfile_source = kerchunk/_version.py
+versionfile_build = kerchunk/_version.py
 tag_prefix = ""
 
 [flake8]
